Added class implementation for GPS Broadcaster
