+++ 2017-2018/ROS_RoboBuggy/src/robobuggy/src/gps/GPS_Broadcaster.cpp
@@ -1,19 +1,8 @@
-#include "ros/ros.h"
-#include "robobuggy/GPS.h"
+#include "gps/GPS_Broadcaster.h"
 
-std::string NODE_NAME = "GPS_Broadcaster";
-
-int main(int argc, char **argv)
+GPS_Broadcaster::GPS_Broadcaster()
 {
-    ros::init(argc, argv, NODE_NAME);
-
-    ros::NodeHandle nh;
-
-    ros::Publisher gps_pub = nh.advertise<robobuggy::GPS>("GPS", 1000);
-
-    float start_lat = 0.0;
-    float start_lon = 0.0;
-    float increment = 0.00001;
+    gps_pub = nh.advertise<robobuggy::GPS>("GPS", 1000);
 
     // Get parameters start lat and start lon
     if (!nh.getParam(NODE_NAME + "/start_latitude", start_lat))
@@ -24,27 +13,20 @@ int main(int argc, char **argv)
     {
         ROS_ERROR("Couldn't retrieve start longitude parameter from parameter server");
     }
-
-    ros::Rate loop_rate(10);
-
-    // Infinitely spin sending dummy GPS messages
-    while(ros::ok()) 
-    {
-        robobuggy::GPS msg;
-
-        int random = rand() % 10;
-        float new_lat = start_lat + increment * random;
-        float new_lon = start_lon + increment * random;
-
-        msg.Lat_deg = new_lat;
-        msg.Long_deg = new_lon;
-
-        gps_pub.publish(msg);
-
-        ros::spinOnce();
-
-        loop_rate.sleep();
-    }
-
-    return 0;
 }
+
+void GPS_Broadcaster::publish_spoof_message()
+{
+    robobuggy::GPS msg;
+
+    int random = rand() % 10;
+    float new_lat = start_lat + increment * random;
+    float new_lon = start_lon + increment * random;
+
+    msg.Lat_deg = new_lat;
+    msg.Long_deg = new_lon;
+
+    gps_pub.publish(msg);
+}
+
+
